LPS-55233 Move methods to JournalFolder services (no logic changes)
