+++ portal-impl/src/com/liferay/portlet/journal/service/impl/JournalFolderServiceImpl.java
@@ -16,17 +16,22 @@ package com.liferay.portlet.journal.service.impl;
 
 import com.liferay.portal.kernel.dao.orm.QueryDefinition;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.security.permission.ActionKeys;
+import com.liferay.portal.security.permission.PermissionChecker;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.util.PropsValues;
+import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
+import com.liferay.portlet.dynamicdatamapping.service.permission.DDMStructurePermission;
 import com.liferay.portlet.journal.model.JournalArticle;
 import com.liferay.portlet.journal.model.JournalFolder;
 import com.liferay.portlet.journal.service.base.JournalFolderServiceBaseImpl;
 import com.liferay.portlet.journal.service.permission.JournalFolderPermission;
 
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 
 /**
@@ -223,6 +228,16 @@ public class JournalFolderServiceImpl extends JournalFolderServiceBaseImpl {
 		}
 	}
 
+	@Override
+	public List<DDMStructure> getJournalFolderStructures(
+			long[] groupIds, long journalFolderId, int restrictionType)
+		throws PortalException {
+
+		return filterStructures(
+			ddmStructureLocalService.getJournalFolderStructures(
+				groupIds, journalFolderId, restrictionType));
+	}
+
 	/**
 	 * @deprecated As of 7.0.0, replaced by {@link #getSubfolderIds(List, long,
 	 *             long, boolean)}
@@ -365,6 +380,28 @@ public class JournalFolderServiceImpl extends JournalFolderServiceBaseImpl {
 			getUserId(), groupId, folderId, parentFolderId, name, description,
 			ddmStructureIds, restrictionType, mergeWithParentFolder,
 			serviceContext);
+	}
+
+	protected List<DDMStructure> filterStructures(List<DDMStructure> structures)
+		throws PortalException {
+
+		PermissionChecker permissionChecker = getPermissionChecker();
+
+		structures = ListUtil.copy(structures);
+
+		Iterator<DDMStructure> itr = structures.iterator();
+
+		while (itr.hasNext()) {
+			DDMStructure structure = itr.next();
+
+			if (!DDMStructurePermission.contains(
+					permissionChecker, structure, ActionKeys.VIEW)) {
+
+				itr.remove();
+			}
+		}
+
+		return structures;
 	}
 
 }
